TRUNK-6206: Global property specific privileges

+++ api/src/main/java/org/openmrs/api/impl/AdministrationServiceImpl.java
@@ -18,6 +18,7 @@ import java.util.Calendar;
 import java.util.Collection;
 import java.util.HashMap;
 import java.util.HashSet;
+import java.util.Iterator;
 import java.util.LinkedHashMap;
 import java.util.LinkedHashSet;
 import java.util.List;
@@ -33,6 +34,7 @@ import org.openmrs.ConceptSource;
 import org.openmrs.GlobalProperty;
 import org.openmrs.ImplementationId;
 import org.openmrs.OpenmrsObject;
+import org.openmrs.Privilege;
 import org.openmrs.User;
 import org.openmrs.api.APIException;
 import org.openmrs.api.AdministrationService;
@@ -150,7 +152,110 @@ public class AdministrationServiceImpl extends BaseOpenmrsService implements Adm
 			return null;
 		}
 		
-		return dao.getGlobalProperty(propertyName);
+		String globalProperty = dao.getGlobalProperty(propertyName);
+		if (StringUtils.isBlank(globalProperty)) {
+			return null;
+		} else {
+			GlobalProperty property = dao.getGlobalPropertyObject(propertyName);
+			if (canViewGlobalProperty(property, null)) {
+				return globalProperty;
+			} else {
+				throw new APIException("GlobalProperty.error.privilege.required.view", new Object[] {
+					property.getViewPrivilege().getPrivilege(), propertyName });
+			}
+		}
+	}
+	
+	/**
+	 * @see org.openmrs.api.AdministrationService#canViewGlobalProperty(org.openmrs.GlobalProperty, org.openmrs.User)
+	 */
+	@Override
+	public boolean canViewGlobalProperty(GlobalProperty property, User user) {
+		if (property == null) {
+			throw new IllegalArgumentException("Global property argument can not be null");
+		}
+		
+		// if user is not specified, then use authenticated user from context by default
+		if (user == null) {
+			user = Context.getAuthenticatedUser();
+		}
+		
+		// since we restrict by view privilege, if it does not exist, then anyone is allowed to view the global property
+		if (property.getViewPrivilege() == null) {
+			return Boolean.TRUE;
+		}
+		
+		return user.hasPrivilege(property.getViewPrivilege().getPrivilege());
+	}
+	
+	/**
+	 * @see org.openmrs.api.AdministrationService#canDeleteGlobalProperty(org.openmrs.GlobalProperty, org.openmrs.User)
+	 */
+	@Override
+	public boolean canDeleteGlobalProperty(GlobalProperty property, User user) {
+		if (property == null) {
+			throw new IllegalArgumentException("Global property argument can not be null");
+		}
+		
+		// if user is not specified, then use authenticated user from context by default
+		if (user == null) {
+			user = Context.getAuthenticatedUser();
+		}
+		
+		// since we restrict by delete privilege, if it does not exist, then anyone is allowed to delete the global property
+		if (property.getDeletePrivilege() == null) {
+			return Boolean.TRUE;
+		}
+		
+		return user.hasPrivilege(property.getDeletePrivilege().getPrivilege());
+	}
+	
+	/**
+	 * @see org.openmrs.api.AdministrationService#canEditGlobalProperty(org.openmrs.GlobalProperty, org.openmrs.User)
+	 */
+	@Override
+	public boolean canEditGlobalProperty(GlobalProperty property, User user) {
+		if (property == null) {
+			throw new IllegalArgumentException("Global property argument can not be null");
+		}
+		
+		// if user is not specified, then use authenticated user from context by default
+		if (user == null) {
+			user = Context.getAuthenticatedUser();
+		}
+		
+		// since we restrict by edit privilege, if it does not exist, then anyone is allowed to edit the global property
+		if (property.getEditPrivilege() == null) {
+			return Boolean.TRUE;
+		}
+		
+		return user.hasPrivilege(property.getEditPrivilege().getPrivilege());
+	}
+	
+	/**
+	 * @see org.openmrs.api.AdministrationService#filterGlobalPropertiesByViewPermissions(java.util.List,
+	 *      org.openmrs.User)
+	 */
+	@Override
+	@Transactional(readOnly = true)
+	public List<GlobalProperty> filterGlobalPropertiesByViewPermissions(List<GlobalProperty> properties, User user) {
+		if (properties != null) {
+			// if user is not specified then use authenticated user from context by default
+			if (user == null) {
+				user = Context.getAuthenticatedUser();
+			}
+			for (Iterator<GlobalProperty> iterator = properties.iterator(); iterator.hasNext();) {
+				GlobalProperty property = iterator.next();
+				// determine whether it's need to include this globalProperty into result or not
+				// as it can be not accessed by current user due to permissions lack
+				Privilege vp = property.getViewPrivilege();
+				if (vp != null && !user.hasPrivilege(vp.getPrivilege())) {
+					// exclude this globalProperty from result
+					iterator.remove();
+				}
+			}
+		}
+		return properties;
 	}
 	
 	/**
@@ -201,6 +306,13 @@ public class AdministrationServiceImpl extends BaseOpenmrsService implements Adm
 		if (gp == null) {
 			throw new IllegalStateException("Global property with the given propertyName does not exist" + propertyName);
 		}
+		
+		// if authenticated user is not supposed to edit globalProperty
+		if (!canEditGlobalProperty(gp, null)) {
+			throw new APIException("GlobalProperty.error.privilege.required.edit", new Object[] {
+				gp.getEditPrivilege().getPrivilege(), propertyName });
+		}
+		
 		gp.setPropertyValue(propertyValue);
 		dao.saveGlobalProperty(gp);
 	}
@@ -211,7 +323,8 @@ public class AdministrationServiceImpl extends BaseOpenmrsService implements Adm
 	@Override
 	@Transactional(readOnly = true)
 	public List<GlobalProperty> getAllGlobalProperties() throws APIException {
-		return dao.getAllGlobalProperties();
+		return Context.getAdministrationService()
+			.filterGlobalPropertiesByViewPermissions(dao.getAllGlobalProperties(), null);
 	}
 	
 	/**
@@ -220,7 +333,8 @@ public class AdministrationServiceImpl extends BaseOpenmrsService implements Adm
 	@Override
 	@Transactional(readOnly = true)
 	public List<GlobalProperty> getGlobalPropertiesByPrefix(String prefix) {
-		return dao.getGlobalPropertiesByPrefix(prefix);
+		return Context.getAdministrationService()
+			.filterGlobalPropertiesByViewPermissions(dao.getGlobalPropertiesByPrefix(prefix), null);
 	}
 	
 	/**
@@ -229,7 +343,8 @@ public class AdministrationServiceImpl extends BaseOpenmrsService implements Adm
 	@Override
 	@Transactional(readOnly = true)
 	public List<GlobalProperty> getGlobalPropertiesBySuffix(String suffix) {
-		return dao.getGlobalPropertiesBySuffix(suffix);
+		return Context.getAdministrationService()
+			.filterGlobalPropertiesByViewPermissions(dao.getGlobalPropertiesBySuffix(suffix), null);
 	}
 	
 	/**
@@ -237,6 +352,11 @@ public class AdministrationServiceImpl extends BaseOpenmrsService implements Adm
 	 */
 	@Override
 	public void purgeGlobalProperty(GlobalProperty globalProperty) throws APIException {
+		// if authenticated user is not supposed to purge globalProperty
+		if (!canDeleteGlobalProperty(globalProperty, null)) {
+			throw new APIException("GlobalProperty.error.privilege.required.purge", new Object[] {
+				globalProperty.getDeletePrivilege().getPrivilege(), globalProperty.getProperty() });
+		}
 		notifyGlobalPropertyDelete(globalProperty.getProperty());
 		dao.deleteGlobalProperty(globalProperty);
 	}
@@ -267,6 +387,13 @@ public class AdministrationServiceImpl extends BaseOpenmrsService implements Adm
 	public GlobalProperty saveGlobalProperty(GlobalProperty gp) throws APIException {
 		// only try to save it if the global property has a key
 		if (gp.getProperty() != null && gp.getProperty().length() > 0) {
+			
+			// if authenticated user is not supposed to save globalProperty
+			if (!canEditGlobalProperty(gp, null)) {
+				throw new APIException("GlobalProperty.error.privilege.required.edit", new Object[] {
+					gp.getEditPrivilege().getPrivilege(), gp.getProperty() });
+			}
+			
 			if (gp.getProperty().equals(OpenmrsConstants.GLOBAL_PROPERTY_LOCALE_ALLOWED_LIST)) {
 				if (gp.getPropertyValue() != null) {
 					List<Locale> localeList = new ArrayList<>();
@@ -621,7 +748,15 @@ public class AdministrationServiceImpl extends BaseOpenmrsService implements Adm
 	@Override
 	@Transactional(readOnly = true)
 	public GlobalProperty getGlobalPropertyByUuid(String uuid) {
-		return dao.getGlobalPropertyByUuid(uuid);
+		GlobalProperty gp =  dao.getGlobalPropertyByUuid(uuid);
+		if (gp == null) {
+			return null;
+		} else if (canViewGlobalProperty(gp, null)) {
+			return gp;
+		} else {
+			throw new APIException("GlobalProperty.error.privilege.required.view", new Object[] {
+				gp.getViewPrivilege().getPrivilege(), gp.getProperty() });
+		}
 	}
 	
 	/**
